// src/Shared/scripts/Enums/50 - scWindowType.ts
export enum ScWindowType {
    Unknown = 0,

    AccessViewer = 50807454,
    Archive = 50572312,
    ContentEditor = 50570457,
    ControlPanel = 50180952,
    Debug = 50053957,
    Desktop = 50670576,
    DomainManager = 50904399,
    EmailExperienceManager = 50111451,
    ExperienceAnalytics = 50849816,
    ExperienceEditor_Edit = 50739548,
    ExperienceEditor_Normal = 50417587,
    ExperienceEditor_Preview = 50495346,
    ExperienceOptimization = 50860380,
    ExperienceProfile = 50307227,
    FallBack = 50657295,
    FederatedExperienceManager = 50999497,
    Forms = 50698620,
    InstallationWizard = 50201337,
    InstalledLicenses = 50979897,
    KeyboardMap = 50683086,
    Launchpad = 50206433,
    LicenseDetails = 50268242,
    ListManager = 50595779,
    LoginPage = 50010467,
    LogViewer = 50500367,
    MarketingAutomation = 50921088,
    MarketingControlPanel = 50528944,
    MediaLibrary = 50351702,
    PackageDesigner = 50363323,
    PathAnalyzer = 50854514,
    Publish = 50797841,
    RecycleBin = 50374978,
    RoleManager = 50017991,
    Run = 50340856,
    ScanForBrokenLinks = 50599314,
    SecurityEditor = 50613652,
    TemplateManager = 50415542,
    UpdateCenter = 50109795,
    UserManager = 50660151,
    Workbox = 50799485,
    InstallerBuildPackage = 50625341,
    JqueryModalDialogs = 50130845,
    ScContentProxy
}